Python/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.3
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.75
multidict==6.0.5
numpy==1.26.4
openai==1.32.0
orjson==3.10.3
packaging==23.2
pydantic==2.7.3
pydantic_core==2.18.4
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
yarl==1.9.4
